refactor: remove dyn_pool module
