+++ src/bin/usage.rs
@@ -1,4 +1,4 @@
-use xanadu::ecs::dyn_pool::{Mut, World};
+use xanadu::ecs::{Mut, World};
 
 #[repr(C)]
 #[derive(Debug, Default, Clone, Copy, bytemuck::Pod, bytemuck::Zeroable, PartialEq)]
