--- src/ecs/dyn_pool.rs
@@ -1,13 +1,0 @@
-mod any_component_array;
-mod component;
-mod query;
-mod system;
-mod typed_component_array;
-mod world;
-
-pub use any_component_array::*;
-pub use component::*;
-pub use query::*;
-pub use system::*;
-pub use typed_component_array::*;
-pub use world::*;
